debianize: Add kernel module blacklist to librtlsdr0 package
This linux video driver blacklist was introduced in librtlsdr0 with
version 0.5.3-2 and later removed with version 0.5.3-10 [1].

The removal was accompanied by the (re-)activation of the automatic
detachment of the the kernel driver (DETACH_KERNEL_DRIVER=ON) [2] for
debian package builds. Since the debianize_scripts in this fork
doesn't enforce specific compile time options (caused by the lack of
a debian/rules file), it should be considered that this option isn't set.
Thus the installation of the blacklist is added in this commit.

[1]: "* Have librtlsdr0 also install a blacklist for linux video drivers"
     "  * Enable DETACH_KERNEL_DRIVER (Closes: 823022)"
     "* remove rtl-sdr-blacklist.conf on upgrade. Thanks Bob! (Closes: #829517)",
     https://anonscm.debian.org/cgit/users/bottoms/pkg-rtl-sdr.git/tree/debian/changelog

[2]: "lib: don't detach kernel driver by default"
     https://github.com/steve-m/librtlsdr/commit/3e17ef2b55d80237da9766f3464ae784448a23e6

Fixes librtlsdr/librtlsdr#19.